# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: List a user's in-app notifications
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notifications, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "400":
          description: Missing user id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: Mark one notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUserRequest"
      responses:
        "204":
          description: Notification marked as read.
        "400":
          description: Invalid notification request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      summary: Mark all notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUserRequest"
      responses:
        "204":
          description: Notifications marked as read.
        "400":
          description: Invalid notification request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /contracts:
    get:
      tags: [Contracts]
      operationId: listContracts
      summary: List contract documents, optionally with a user's read status
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Contract documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractDocument"
  /contracts/{slug}/read:
    post:
      tags: [Contracts]
      operationId: markContractRead
      summary: Mark a contract document as read by a user
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkContractReadRequest"
      responses:
        "200":
          description: Updated read progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDocument"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Document not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contracts/acceptance:
    get:
      tags: [Contracts]
      operationId: getContractAcceptance
      summary: Get a user's contract acceptance record
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Acceptance record, or null if not yet accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractAcceptanceResponse"
    post:
      tags: [Contracts]
      operationId: acceptContracts
      summary: Record a user's final contract acceptance with a server timestamp
      description: |
        Fails with 409 if the user has not yet marked all contract documents
        as read.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptContractsRequest"
      responses:
        "200":
          description: Acceptance recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractAcceptance"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Not all contract documents have been read yet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Notification:
      type: object
      required: [id, userId, title, body, createdAt, read]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
        read:
          type: boolean
    NotificationUserRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          minLength: 1
        name:
          type: string
        email:
          type: string
    ContractDocument:
      type: object
      required: [slug, title, read]
      properties:
        slug:
          type: string
        title:
          type: string
        objectPath:
          type: string
          nullable: true
        read:
          type: boolean
        readAt:
          type: string
          nullable: true
    MarkContractReadRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          minLength: 1
    ContractAcceptance:
      type: object
      required: [userId, emergencyContactName, emergencyContactPhone, acceptedAt]
      properties:
        userId:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        acceptedAt:
          type: string
        guardianName:
          type: string
          nullable: true
        guardianRelationship:
          type: string
          nullable: true
        minorDataConsentAt:
          type: string
          nullable: true
          description: |
            Set only when the member accepted while under 18 — the timestamp
            of the separate consent to process the minor's personal data,
            distinct from acceptedAt.
    ContractAcceptanceResponse:
      type: object
      properties:
        acceptance:
          $ref: "#/components/schemas/ContractAcceptance"
          nullable: true
    AcceptContractsRequest:
      type: object
      required: [userId, emergencyContactName, emergencyContactPhone]
      properties:
        userId:
          type: string
          minLength: 1
        emergencyContactName:
          type: string
          minLength: 1
        emergencyContactPhone:
          type: string
          minLength: 1
        guardianName:
          type: string
          minLength: 1
          description: |
            Required when the member is under 18. The guardian's acceptance
            stands in for the minor's own signature.
        guardianRelationship:
          type: string
          minLength: 1
          description: Optional, e.g. "Madre", "Padre", "Tutor legal".
        minorDataConsent:
          type: boolean
          description: |
            Required to be true when the member is under 18: an explicit,
            separate consent to process the minor's personal data for this
            app, distinct from accepting the box's contract. The server
            records its own timestamp (minorDataConsentAt).
...